requirements.txt
awscli==1.6.2
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
bcdoc==0.12.2
boto==2.33.0
botocore==0.73.0
certifi==2015.4.28
colorama==0.2.5
cycler==0.10.0
decorator==4.0.9
defusedxml==0.4.1
Django==1.8.8
django-ajax-search==1.5.1
django-braces==1.9.0
django-el-pagination==2.1.2
django-haystack==2.4.1
django-likert-field==0.2.0
django-model-utils==2.5
django-tastypie==0.13.3
docutils==0.12
freeze==1.0.10
functools32==3.2.3.post2
genson==0.1.0
GraphLab-Create==1.10.1
GraphLab-Create-License==1.10.1
jmespath==0.5.0
jsonschema==2.5.1
lxml==3.6.0
matplotlib==1.5.1
multipledispatch==0.4.8
numpy==1.11.1
pandas==0.18.1
prettytable==0.7.2
pyasn1==0.1.9
pyparsing==2.1.5
python-dateutil==2.5.3
python-mimeparse==1.5.2
pytz==2016.4
requests==2.9.1
rsa==3.1.2
scipy==0.17.1
simplejson==3.8.2
singledispatch==3.4.0.3
six==1.10.0
sseclient==0.0.8
tornado==4.3
Whoosh==2.7.4
